requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
easyocr==1.7.1
pdfplumber==0.10.2
spacy==3.7.2
python-Levenshtein==0.23.0 